Moved settings to own namespace
