+++ video-takeoff-control/CompetitionNameModal.xaml.cs
@@ -11,6 +11,7 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
+using video_takeoff_control.settings;
 
 namespace video_takeoff_control
 {
