+++ video-takeoff-control/video-source/SimpleHttpVideoSource.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using video_takeoff_control.logging;
+using video_takeoff_control.settings;
 
 namespace video_takeoff_control.video_source
 {
